almost finish the admin side of the application

+++ FifthAssignment.Core.Application/Models/CreditCardModels/SaveCreditCardModel.cs
@@ -3,7 +3,8 @@ namespace FifthAssignment.Core.Application.Models.CreditCardModels
 {
 	public class SaveCreditCardModel
 	{
-		public double Amount { get; set; }
+
+		public decimal Amount { get; set; }
 		public string IdentifierNumber { get; set; }
 		public int CreditLimit { get; set; }
 		public string UserId { get; set; }
